# ddtrace/tracer.Measured(): Cache a global instance (save 1 alloc/call)